nutriscan-ai/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.51.0
httpx==0.27.2
pydantic>=2.9.2,<3
python-multipart==0.0.12
aiofiles==24.1.0
email-validator==2.2.0
